separated rest service into parts
