+++ src/main/scala/io/andrewsmith/website/Main.scala
@@ -2,7 +2,7 @@ package io.andrewsmith.website
 
 import cats.effect._
 import cats.implicits._
-import io.andrewsmith.website.services.{FileService, ResourceService, RestService, ViewService}
+import io.andrewsmith.website.services._
 import org.http4s.HttpApp
 import org.http4s.implicits._
 import org.http4s.server.Router
@@ -12,7 +12,7 @@ import org.http4s.server.blaze._
 object Main extends IOApp {
   private val app: HttpApp[IO] = Router(
     "/" -> FileService.routes.combineK(ViewService.routes),
-    "/api" -> RestService.routes,
+    "/api" -> BogosortApiService.routes.combineK(MessagesApiService.routes),
     "/assets" -> ResourceService.routes
   ).orNotFound
